prom_order_manager/requirements.txt
requests
python-telegram-bot
python-dotenv
pandas
openpyxl
flask

